Exeption was Improved, also new exceptions were added on the controller.
