+++ src/main/java/com/jalasoft/practice/Main.java
@@ -7,7 +7,7 @@
  * license agreement you entered into with Jalasoft.
  */
 
-package com.jalasoft.convert;
+package com.jalasoft.practice;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.SpringApplication;
