+++ src/main/java/com/jalasoft/practice/controller/constant/ErrorConstant.java
@@ -1,0 +1,19 @@
+/*
+ *  Copyright (c) 2020 Jalasoft.
+ *
+ *  This software is the confidential and proprietary information of Jalasoft.
+ *  ("Confidential Information"). You shall not disclose such Confidential
+ *  Information and shall use it only in accordance with the terms of the
+ *  license agreement you entered into with Jalasoft.
+ *
+ */
+
+package com.jalasoft.practice.controller.constant;
+
+/**
+ * @version 1.1
+ * @autor Magdalena
+ */
+public class ErrorConstant {
+    public static final String FILE_ERROR = "Failed to create file";
+}
