+++ src/main/java/com/jalasoft/practice/controller/exception/FileException.java
@@ -1,0 +1,25 @@
+/*
+ *  Copyright (c) 2020 Jalasoft.
+ *
+ *  This software is the confidential and proprietary information of Jalasoft.
+ *  ("Confidential Information"). You shall not disclose such Confidential
+ *  Information and shall use it only in accordance with the terms of the
+ *  license agreement you entered into with Jalasoft.
+ *
+ */
+
+package com.jalasoft.practice.controller.exception;
+
+/**
+ * @version 1.1
+ * @autor Magdalena
+ */
+public class FileException extends Exception {
+    public FileException(String message, Throwable ex) {
+        super(message, ex);
+    }
+
+    public FileException(String message) {
+        super(message);
+    }
+}
