+++ src/main/java/com/jalasoft/practice/controller/response/ErrorResponse.java
@@ -1,0 +1,35 @@
+/*
+ *  Copyright (c) 2020 Jalasoft.
+ *
+ *  This software is the confidential and proprietary information of Jalasoft.
+ *  ("Confidential Information"). You shall not disclose such Confidential
+ *  Information and shall use it only in accordance with the terms of the
+ *  license agreement you entered into with Jalasoft.
+ *
+ */
+
+package com.jalasoft.practice.controller.response;
+
+/**
+ * @version 1.1
+ * @autor Magdalena
+ */
+public class ErrorResponse<T> extends Response {
+
+    String error;
+
+    public ErrorResponse(String error, T status) {
+        super(status);
+        this.error = error;
+    }
+
+    public String getError() {
+        return error;
+    }
+
+    public void setError(String error) {
+        this.error = error;
+    }
+
+
+}
