+++ src/main/java/com/jalasoft/practice/controller/utils/CastMultipartToFile.java
@@ -8,7 +8,7 @@
  *
  */
 
-package com.jalasoft.convert.controller.utils;
+package com.jalasoft.practice.controller.utils;
 
 import org.springframework.web.multipart.MultipartFile;
